+++ src/platform/window_sdl.rs
@@ -234,7 +234,7 @@ impl Window {
         self.window_visible = visible;
     }
 
-    pub fn get_device_pixel_ratio(&self) -> f32 {
+    pub fn get_dpi_scale(&self) -> f32 {
         self.sdl_window.drawable_size().0 as f32 / self.sdl_window.size().0 as f32
     }